load node with another docker image
Added a new submenu in the PC node. When not empty, it loads the given
docker image instead of the default imunes/vroot one.
Still needs testing and inserting this menu to all the other nodes, but
for now, this works as expected on my computer (tried it with httpd image).